+++ samplesheets/vueapp/tests/unit/StudyShortcutModal.spec.js
@@ -3,6 +3,7 @@ import {
   projectUuid,
   studyUuid,
   assayUuid,
+  copy,
   waitNT,
   waitRAF
 } from '../testUtils.js'
@@ -20,6 +21,9 @@ const fetchMock = require('fetch-mock-jest')
 // Init data
 let propsData
 const ajaxUrl = '/samplesheets/ajax/study/links/' + studyUuid + '?case=A001'
+
+// Local constants
+const omitPrefix = 'Omitted from IGV: '
 
 describe('StudyShortcutModal.vue', () => {
   function getPropsData () {
@@ -55,10 +59,35 @@ describe('StudyShortcutModal.vue', () => {
     expect(wrapper.find('#sodar-ss-shortcut-modal').exists()).toBe(true)
     expect(wrapper.findAll('.sodar-ss-shortcut-item').length).toBe(3)
     expect(wrapper.findAll('.sodar-ss-shortcut-extra').length).toBe(4)
+    expect(wrapper.findAll('.sodar-ss-shortcut-info').length).toBe(2)
+    expect(wrapper.findAll('.sodar-ss-shortcut-info').at(0).attributes().title).toBe(
+      omitPrefix + '*dragen_evidence.bam')
+    expect(wrapper.findAll('.sodar-ss-shortcut-info').at(1).attributes().title).toBe(
+      omitPrefix + '*cnv.vcf.gz, *ploidy.vcf.gz, *sv.vcf.gz')
   })
 
-  it('renders modal with message', async () => {
-    fetchMock.mock(ajaxUrl, { detail: 'Message' })
+  it('renders modal with empty omit info data', async () => {
+    const copiedShortcuts = copy(studyShortcuts)
+    copiedShortcuts.data.vcf.omit_info = ''
+    fetchMock.mock(ajaxUrl, copiedShortcuts)
+    const wrapper = mount(StudyShortcutModal, {
+      localVue, propsData: propsData
+    })
+    wrapper.vm.showModal({ key: 'case', value: 'A001' })
+    await waitNT(wrapper.vm)
+    await waitRAF()
+
+    expect(fetchMock.called(ajaxUrl)).toBe(true)
+    expect(wrapper.find('#sodar-ss-shortcut-modal').exists()).toBe(true)
+    expect(wrapper.findAll('.sodar-ss-shortcut-item').length).toBe(3)
+    expect(wrapper.findAll('.sodar-ss-shortcut-extra').length).toBe(4)
+    expect(wrapper.findAll('.sodar-ss-shortcut-info').length).toBe(1)
+    expect(wrapper.findAll('.sodar-ss-shortcut-info').at(0).attributes().title).toBe(
+      omitPrefix + '*dragen_evidence.bam')
+  })
+
+  it('renders modal with error message', async () => {
+    fetchMock.mock(ajaxUrl, { error: 'Message' })
     const wrapper = mount(StudyShortcutModal, {
       localVue, propsData: propsData
     })